patchCmdLine.xul: Add user interface to manage the options.
We get fully configurable! This commit adds textboxes to choose what
names to use for the options, and checkboxes to select what options to
use.
The hidden post-processing regexp introduced in a previous commit can
now be configured easily on the right side of the interface.

Misc: a debug function associated with a hidden debug label on the
interface has been added to help the developer. Users should not see it,
and can safely ignore it.